# home/WEB-INF/openapi/zirc-api.yaml
openapi: 3.0.3
info:
  title: ZIRC Line Submission API
  version: 1.0.0
  description: |
    JSON API backing the ZIRC line-submission curator workflow. Three
    aggregates, each with its own field-path PATCH endpoint:

      - line submission       /api/zirc/line-submissions/{zdbID}
      - mutation              /api/zirc/mutations/{mutationId}
      - genotyping assay      /api/zirc/assays/{assayId}

    Writes are per-field: clients send `{path, value}` with a JSON Pointer
    and the server applies the change atomically and writes one audit row.
    Form schemas (JSON Schema + JSON Forms uiSchema) are served from a
    parallel `*/form-schema` endpoint per aggregate.

    Errors are returned as RFC 7807 `application/problem+json`.
servers:
  - url: /action/api/zirc
    description: ZIRC API base path under the Tomcat /action prefix
...
paths:
...
  /line-submissions/{zdbID}:
    parameters:
      - $ref: '#/components/parameters/SubmissionZdbId'
    get:
      tags: [line-submission]
      summary: Fetch a line submission by ZDB ID.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LineSubmissionDTO' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [line-submission]
      summary: Apply a single field-path change. Paths and value types are constrained by the form schema; unknown paths are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FieldUpdate' }
      responses:
        '200':
          description: Updated submission.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LineSubmissionDTO' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /line-submissions/{zdbID}/mutations:
    parameters:
      - $ref: '#/components/parameters/SubmissionZdbId'
    post:
      tags: [mutation]
      summary: Add an empty mutation row to a submission.
      responses:
        '200':
          description: Newly-created mutation (parent submission's mutations list now contains it).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationDTO' }
        '404': { $ref: '#/components/responses/NotFound' }
  /line-submissions/{zdbID}/mutations/{mutationId}:
    parameters:
      - $ref: '#/components/parameters/SubmissionZdbId'
      - $ref: '#/components/parameters/MutationId'
    delete:
      tags: [mutation]
      summary: Remove a mutation from a submission.
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    SubmissionZdbId:
      name: zdbID
      in: path
      required: true
      description: ZFIN active-data ID for a line submission.
      schema:
        type: string
        example: ZDB-LINESUBMISSION-260515-7
    MutationId:
      name: mutationId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: Invalid request (rejected before persistence).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: Aggregate not found by the given id.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    ValidationFailed:
      description: Body failed Jakarta Bean Validation. The problem detail's `errors` extension carries field → message.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
    FieldUpdate:
      type: object
      required: [path]
      description: A single change to one form-schema path. `path` is a JSON Pointer rooted at the aggregate (`/name`, `/acceptance/reasons`, etc.). `value` is whatever the corresponding FieldDescriptor accepts — string, boolean, null, array of strings, or array of objects depending on the field.
      properties:
        path:
          type: string
          example: '/acceptance/reasons'
        value:
          nullable: true
          description: Permitted shape varies by `path`; see the form schema for that aggregate.
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem detail. Returned by ZircApiExceptionHandler for known-bad inputs and missing entities.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties: { type: string }
          description: Field → message map; present on 422 ValidationFailed only.
    LineSubmissionDTO:
      type: object
      properties:
        zdbID: { type: string }
        name: { type: string, nullable: true }
        abbreviation: { type: string, nullable: true }
        previousNames: { type: string, nullable: true }
        singleAllelic: { type: boolean, nullable: true }
        maternalBackground: { type: string, nullable: true }
        paternalBackground: { type: string, nullable: true }
        backgroundChangeable: { type: boolean, nullable: true }
        backgroundChangeConcerns: { type: string, nullable: true }
        unreportedFeaturesDetails: { type: string, nullable: true }
        husbandryInfo: { type: string, nullable: true }
        additionalInfo: { type: string, nullable: true }
        reasons:
          type: array
          items: { type: string }
        reasonsOther: { type: string, nullable: true }
        mutations:
          type: array
          items: { $ref: '#/components/schemas/MutationDTO' }
        draft: { type: boolean }
    MutationDTO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        lineSubmissionId: { type: string }
        sortOrder: { type: integer }
        alleleDesignation: { type: string, nullable: true }
        alleleInZfin: { type: boolean, nullable: true }
        mutationType: { type: string, nullable: true }
        mutationDiscoverer: { type: string, nullable: true }
        mutationInstitution: { type: string, nullable: true }
        mutagenesisStage: { type: string, nullable: true }
        mutagenesisProtocol: { type: string, nullable: true }
        molecularlyCharacterized: { type: boolean, nullable: true }
        homozygousLethal: { type: boolean, nullable: true }
        lethalityStageTypical: { type: string, nullable: true }
        lethalitySpecificTimepoint: { type: string, nullable: true }
        lethalityWindowStart: { type: string, nullable: true }
        lethalityWindowEnd: { type: string, nullable: true }
        lethalityAdditionalInfo: { type: string, nullable: true }
        publications:
          type: array
          items: { type: string }
        assays:
          type: array
          items: { $ref: '#/components/schemas/AssaySummaryDTO' }
    AssaySummaryDTO:
      type: object
      description: Compact row used inside MutationDTO.assays; full fields fetched separately via /assays/{id}.
      properties:
        id: { type: integer, format: int64 }
        sortOrder: { type: integer }
        assayType: { type: string, nullable: true }
...